requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.2.0
certifi==2021.5.30
chardet==3.0.4
discord.py==1.7.3
idna==2.10
install==1.3.4
multidict==5.1.0
python-dotenv==0.19.0
requests==2.23.0
saucenao-api==2.4.0
typing-extensions==3.10.0.2
urllib3==1.25.11
yarl==1.6.3
